@@ -1,15 +1,21 @@
-import 'package:appstructure/ui/dashboard/dashboard_page.dart';
-import 'package:appstructure/ui/home/home.dart';
-import 'package:appstructure/ui/splash/splash.dart';
-import 'package:flutter/material.dart';
-
-
 class Routes {
-
   static const String dashboardScreen = '/dashboardScreen';
   static const String splashScreen = '/splashScreen';
   static const String dailydarshan = '/dailydarshan';
   static const String livedarshan = '/livedarshan';
   static const String newslist = '/newslist';
-
+  static const String ghanshyamvijay = '/ghanshyamvijay';
+  static const String aboutus = '/aboutus';
+  static const String contactus = '/contactus';
+  static const String locations = '/locations';
+  static const String dailyQuotes = '/dailyQuotes';
+  static const String nityaniyamslits = '/nityaniyamslits';
+  static const String newsdetail = '/newsdetail';
+  static const String dailylivedarshan = '/dailylivedarshan';
+  static const String more = '/more';
+  static const String privacypolicy = '/privacypolicy';
+  static const String termsandconditions = '/termsandconditions';
+  static const String underconstruction = '/underconstruction';
+  static const String home = '/home';
+  static const String mapfullscreen = '/mapfullscreen';
 }
