@@ -1,7 +1,7 @@
 import 'package:get/get.dart';
 
 class DashboardController extends GetxController {
-  var tabIndex = 0;
+  var tabIndex = 2;
 
   void changeTabIndex(int index) {
     tabIndex = index;
