@@ -1,17 +1,21 @@
 import 'package:flutter/material.dart';
 
-class CustomColors{
-
-  static const OrangeColor =  Color(0xFFF57700);
-  static const FontColorGray =  Color(0xFF777777);
-  static const FontColorGray2 =  Color(0xFF5B5D62);
-  static const ColorWhite =  Colors.white;
-  static const ColorBlack =  Colors.black;
-  static const ColorBlack2 =  Color(0xFF13161D);
-  static const Gradient1 =  Color(0xFFFF7455);
-  static const Gradient2 =  Color(0xFFED3237);
-  static const Gradient3 =  Color(0xFFFF8240);
-  static const Gradient4 =  Color(0x009E400E);//Daily darshan top view
-  static const Gradient5 =  Color(0x669E400E);//Daily darshan top view
-  static const IconColor =  Color(0xFF373A40);
+class CustomColors {
+  static const OrangeColor = Color(0xFFF57700);
+  static const FontColorGray = Color(0xFF777777);
+  static const FontColorGray2 = Color(0xFF5B5D62);
+  static const ColorWhite = Colors.white;
+  static const ColorBlack = Colors.black;
+  static const ColorBlack2 = Color(0xFF13161D);
+  static const Gradient1 = Color(0xFFFF7455);
+  static const Gradient2 = Color(0xFFED3237);
+  static const Gradient3 = Color(0xFFFF8240);
+  static const Gradient4 = Color(0x009E400E); //Daily darshan top view
+  static const Gradient5 = Color(0x669E400E); //Daily darshan top view
+  static const IconColor = Color(0xFF373A40);
+  static const IndicatorActiveColor = Color(0xFFF57700);
+  static const LinearGradientStart = Color(0xFFFF8240);
+  static const LinearGradientEnd = Color(0xFFFF7455);
+  static const TextColor1 = Color(0xFF373A40);
+  static const GrayColor = Color(0xFFF1F1F1);
 }
