@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 class MaskedTextInputFormatter extends TextInputFormatter {
@@ -14,8 +13,9 @@ class MaskedTextInputFormatter extends TextInputFormatter {
   TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
     if (newValue.text.isNotEmpty) {
       if (newValue.text.length > oldValue.text.length) {
-        if (newValue.text.length > mask.length)
+        if (newValue.text.length > mask.length) {
           return oldValue;
+        }
         if (newValue.text.length < mask.length && mask[newValue.text.length - 1] == separator) {
           return TextEditingValue(
             text: '${oldValue.text}$separator${newValue.text.substring(newValue.text.length - 1)}',
